Ship renamed stylesheet.css in update packages
So far, stylesheet changes are ignored in update packages, what can be
irrelevant, but also a serious issue.  Instead of requiring users to
download the full package, and to grab stylesheet.css from there, we
just ship a renamed stylesheet.  This still requires Webmasters to
manually "merge" the contents, but it's a small help.